% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_h5ad.R
\name{Cells.RawH5ad}
\alias{Cells.RawH5ad}
\title{Get Cell Names}
\usage{
\method{Cells}{RawH5ad}(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Arguments passed to other methods}
}
\description{
Get Cell Names
}
\seealso{
\code{\link[SeuratObject:Cells]{SeuratObject::Cells}}
}
\concept{objects}
\concept{spatial}
