% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_h5ad.R
\name{GetImage.RawH5ad}
\alias{GetImage.RawH5ad}
\title{Get Image Data}
\usage{
\method{GetImage}{RawH5ad}(object, mode = c("grob", "raster", "plotly", "raw"), ...)
}
\arguments{
\item{object}{An object}

\item{mode}{How to return the image; should accept one of \dQuote{grob},
\dQuote{raster}, \dQuote{plotly}, or \dQuote{raw}}

\item{...}{Arguments passed to other methods}
}
\description{
Get Image Data
}
\seealso{
\code{\link[SeuratObject:GetImage]{SeuratObject::GetImage}}
}
\concept{objects}
\concept{spatial}
