% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_h5ad.R
\name{GetTissueCoordinates.RawH5ad}
\alias{GetTissueCoordinates.RawH5ad}
\title{Get Tissue Coordinates}
\usage{
\method{GetTissueCoordinates}{RawH5ad}(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}
}
\description{
Get Tissue Coordinates
}
\seealso{
\code{\link[SeuratObject:GetTissueCoordinates]{SeuratObject::GetTissueCoordinates}}
}
\concept{objects}
\concept{spatial}
