% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_h5ad.R
\name{h5ad_to_seurat}
\alias{h5ad_to_seurat}
\title{Convert H5AD to Seurat V5}
\usage{
h5ad_to_seurat(file_path)
}
\arguments{
\item{file_path}{Path to .h5ad file}
}
\value{
Seurat object
}
\description{
Convert H5AD to Seurat V5
}
