% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_h5ad.R
\name{Radius.RawH5ad}
\alias{Radius.RawH5ad}
\title{Get Spot Radius}
\usage{
\method{Radius}{RawH5ad}(object, ...)
}
\arguments{
\item{object}{An image object}

\item{...}{Arguments passed to other methods}
}
\description{
Get Spot Radius
}
\seealso{
\code{\link[SeuratObject:Radius]{SeuratObject::Radius}}
}
\concept{objects}
\concept{spatial}
