% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_h5ad.R
\name{RenameCells.RawH5ad}
\alias{RenameCells.RawH5ad}
\alias{RenameCells.VisiumV1}
\title{Rename Cells in an Object}
\usage{
\method{RenameCells}{RawH5ad}(object, new.names = NULL, ...)

\method{RenameCells}{VisiumV1}(object, new.names = NULL, ...)
}
\arguments{
\item{object}{An object}

\item{new.names}{vector of new cell names}

\item{...}{Arguments passed to other methods}
}
\description{
Rename Cells in an Object
}
\seealso{
\code{\link[SeuratObject:RenameCells]{SeuratObject::RenameCells}}
}
\concept{objects}
