% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_h5ad.R
\name{seurat_to_h5ad}
\alias{seurat_to_h5ad}
\title{Export Seurat V5 to H5AD}
\usage{
seurat_to_h5ad(
  seurat_obj,
  file_path,
  compression_level = 4,
  write_raw_slot = TRUE
)
}
\arguments{
\item{seurat_obj}{Seurat Object}

\item{file_path}{Output path}

\item{compression_level}{0-9}

\item{write_raw_slot}{boolean}
}
\description{
Export Seurat V5 to H5AD
}
